requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
gql==3.4.1
graphql-core==3.2.3
idna==3.4
multidict==6.0.4
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
